requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
Django==5.2.7
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.0
joblib==1.5.2
lazy_loader==0.4
magic-filter==1.0.12
multidict==6.7.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
propcache==0.4.1
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.2.1
python-telegram-bot==22.5
PyWavelets==1.9.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tifffile==2025.10.16
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
